+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - disallowed implicit conversions between strings and other types
     - corrected related bugs
-- arrays
+- arrays, array indexing
 
 
 ## 0.4